requirements.txt
# Data Ingestion
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0

# Data Processing
pandas>=2.1.0
pyarrow>=14.0.0
duckdb>=0.9.0

# Apache Iceberg
pyiceberg>=0.5.0

# ClickHouse
clickhouse-connect>=0.6.0
clickhouse-driver>=0.2.6

# dbt
dbt-core>=1.7.0
dbt-duckdb>=1.7.0
dbt-clickhouse>=1.7.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Optional: Jupyter for exploration
jupyter>=1.0.0
notebook>=7.0.0

# Optional: Visualization
plotly>=5.17.0
dash>=2.14.0

# Optional: API framework (for exposing data)
fastapi>=0.104.0
uvicorn>=0.24.0
